# CITFAC-openapi.yaml
openapi: 3.0.3
info:
  title: Città Facile
  description: Insieme di API esposte dal BFF di Città Facile (nativamente o attraverso proxy di API esistenti)
  contact:
    email: maurizio.peisino@csi.it
  version: 1.0.0
servers:
  - url: 'https://localhost'
...
paths:
...
  /api-public/configComune/fromtenant:
    get:
      operationId: getComune
      summary: Get comune
      description: Get all comune information
      tags:
        - Public | Comune
      responses:
        '200':
          description: The request has been fulfilled correctly
          content:
            application/json:
              schema:
                type: object
                properties:
                  nome:
                    type: string
                    example: comune_Test
                  codiceImmagine:
                    type: string
                    example: aaaa
                  regione:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      nome:
                        type: string
                        example: Piemonte
                      link:
                        type: string
                        example: www.google.it
                  tenant:
                    type: string
                    example: comune.it
                  firebase:
                    type: object
                    properties:
                      apiKey:
                        type: string
                        example: TEST
                      appId:
                        type: string
                        example: TEST
                      authDomain:
                        type: string
                        example: TEST
                      databaseURL:
                        type: string
                        example: TEST
                      messagingSenderId:
                        type: string
                        example: TEST
                      projectId:
                        type: string
                        example: TEST
                      storageBucket:
                        type: string
                        nullable: true
                      vapidKey:
                        type: string
                        example: TEST
                  assets:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        codice:
                          type: string
                          example: '12345'
                        nome:
                          type: string
                          example: asset1
                        descrizione:
                          type: string
                          example: asset di prova
                  homePage:
                    type: string
                    example: www.google.com
                  codiceEnte:
                    type: string
                    example: c_123
        '204':
          description: No Content
  /api-public/configComune/menu:
    get:
      operationId: getMenuComune
      summary: Get menu
      description: Get the comune menu
      tags:
        - Public | Comune
      responses:
        '200':
          description: Tenant menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '404':
          $ref: '#/components/responses/404'
  /api-public/configComune/servizi:
    get:
      operationId: getServiceComune
      summary: Get service
      description: Get the comune services
      tags:
        - Public | Comune
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Servizio'
        '302':
          description: redirect to cms/servizi_api
  /api-public/configComune/dettaglioServizi:
    get:
      operationId: getDettaglioservizi
      summary: Get dettaglio servizi
      description: Get the dettaglio servizi comune
      tags:
        - Public | Comune
      responses:
        '200':
          description: List of services details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DettaglioServizio'
        '403':
          $ref: '#/components/responses/404'
        '404':
          $ref: '#/components/responses/404'
...
  '/api-admin/tenant/{tenant-id}/service/{service-id}':
    parameters:
      - $ref: '#/components/parameters/idComune'
      - $ref: '#/components/parameters/idServizio'
    get:
      operationId: getService
      summary: Get service
      description: Get single service for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: Tenant service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Servizio'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    put:
      operationId: putService
      summary: Update service
      description: Update single service for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: Updated tenant service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Servizio'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    delete:
      operationId: deleteService
      summary: Delete service
      description: Delete single service for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: Service is deleted
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
...
  '/api-admin/tenant/{tenant-id}/menu':
    parameters:
      - $ref: '#/components/parameters/idComune'
    get:
      operationId: getMenu
      summary: Get menu
      description: Get menu for single tenant
      tags:
        - Administration | Menu
      responses:
        '200':
          description: Tenant menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '404':
          $ref: '#/components/responses/404'
    post:
      operationId: addMenu
      summary: Add menu
      description: Add new menu for single tenant
      tags:
        - Administration | Menu
      responses:
        '201':
          description: Create tenant menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    put:
      operationId: putMenu
      summary: Update menu
      description: Update menu for single tenant
      tags:
        - Administration | Menu
      responses:
        '200':
          description: Updated tenante menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
...
components:
  parameters:
    idComune:
      name: tenant-id
      in: path
      required: true
      schema:
        type: integer
        default: 0
    idServizio:
      name: service-id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    '400':
      description: Bad Request. Body or parameters validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
...
    '403':
      description: Forbidden. User have no access to resource or tenant is not corresponding
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    '404':
      description: Not Found. Cannot find entity key or tenant id related to the entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
...
  schemas:
    Errore:
      type: object
      properties:
        status:
          type: integer
          format: int32
        title:
          type: string
          example: Messaggio di errore
        detail:
          type: array
          nullable: true
          items:
            type: string
            example: 'Nome campo: Dettaglio messaggio di errore'
...
    Servizio:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Nome servizio
        descrizione:
          type: string
          example: Descrizione servizio
        categoriaServizio:
          type: object
          properties:
            id:
              type: integer
              example: 1
            nome:
              type: string
              example: Categoria servizio
        linkScheda:
          type: string
          format: uri
          description: link alla scheda descrittiva del servizio
        linkServizio:
          type: string
          format: uri
          description: link del servizio
        servizioCittaFacile:
          type: boolean
          description: indica se il servizio viene lanciato da città facile
          example: true
        servizioAttivo:
          type: boolean
          description: indica se il servizio è attivo oppure in manutenzione
          example: true
        messaggioManutenzione:
          type: string
          nullable: true
          description: messaggio da mostrare all'utente quando il servizio risulta non attivo
          example: Il servizio non è disponibile per attività di manutenzione straordinaria
        notificabile:
          type: boolean
          example: false
        codiceNotifiche:
          type: string
          nullable: true
        assistenza:
          type: boolean
          example: false
        codice_assistenza:
          type: string
          nullable: true
    DettaglioServizio:
      title: Dettaglio Servizio
      type: object
      properties:
        id:
          type: integer
          example: 1
        titolo:
          type: string
        idServizio:
          type: integer
          description: l'id della servizio proveniente da CMS
          example: 1
        moduloMoon:
          type: string
        codiceNotifiche:
          type: string
          description: il tag con cui recuperare le notifiche relative al servizio (quando è notificabile)
        codiceAssistenza:
          type: string
          description: il codice da utilizzare nel form di assistenza
        idFaq:
          type: integer
          description: l'id della catagoria delle FAQ da CRM
          example: 1
        idBot:
          type: integer
          example: 1
        asset:
          type: string
          example: 'null'
...
    Menu:
      type: object
      properties:
        main:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
        menu-argomenti:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
        contatta-comune:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
    MenuItem:
      type: object
      properties:
        title:
          type: string
          description: label della voce di menù
        url:
          type: string
          format: uri
          description: link della voce di menù
        attributes:
          type: object
          properties:
            data-element:
              type: string
            class:
              type: array
              description: elenco delle classi CSS
              items:
                type: string
        is_parent:
          type: boolean
          description: indica la presenza o meno di un ulteriore livello di menù (usato solo per il menù footer)
        children:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
      example:
        title: Contatta il comune
        url: ''
        attributes:
          data-element: ''
          class: []
        is_parent: true
        children:
          - title: Leggi le domande frequenti
            url: 'https://dev-www-comune-pinerolo-to.portali.csi.it/domande-frequenti'
            attributes:
              class:
                - faq
            is_parent: false
            children: []
...